Update sign.ts
No reason not to return the transaction states as an array. Much easier to use in code than having to use a nested function callback to get this data.
The nested function callbacks `onTrxConfirmed` is useful for other use cases, but for when you just want to get a list of states for your transaction, this is much easier.